pyproject.toml
[project]
name = "assistax"
version = "0.0.1"
description = "A Hardware-Accelerated Reinforcement Learning Benchmark for Assistive Robotics "
readme = "README.md"
authors = [{name = "Leonard Hinckeldey", email = "l.hinckeldey@ed.ac.uk"}]
license = "Apache-2.0"
requires-python = ">=3.11,<3.12"
dependencies = [
    "absl-py==2.3.0",
    "antlr4-python3-runtime==4.9.3",
    "attrs==25.3.0",
    "blinker==1.9.0",
    "brax==0.10.4",
    "chex==0.1.89",
    "click==8.2.1",
    "cloudpickle==3.1.1",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "decorator==5.2.1",
    "distrax==0.1.5",
    "dm-env==1.6",
    "dm-tree==0.1.9",
    "dotmap==1.3.30",
    "etils==1.12.2",
    "farama-notifications==0.0.4",
    "flashbax==0.1.3",
    "flask==3.1.1",
    "flask-cors==6.0.1",
    "flax==0.10.2",
    "fonttools==4.58.4",
    "fsspec==2025.5.1",
    "gast==0.6.0",
    "glfw==2.9.0",
    "grpcio==1.73.1",
    "gym==0.26.2",
    "gym-notices==0.0.8",
    "gymnasium==1.2.0",
    "gymnax==0.0.6",
    "humanize==4.12.3",
    "hydra-core==1.3.2",
    "importlib-resources==6.5.2",
    "itsdangerous==2.2.0",
    "jax[cuda12]==0.4.37",
    "jaxmarl==0.1.0",
    "jaxtyping==0.3.2",
    "jinja2==3.1.6",
    "kiwisolver==1.4.8",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.3",
    "mdurl==0.1.2",
    "ml-collections==1.1.0",
    "ml-dtypes==0.5.1",
    "msgpack==1.1.1",
    "mujoco==3.2.6",
    "mujoco-mjx==3.2.6",
    "nest-asyncio==1.6.0",
    "numpy==1.26.4",
    "omegaconf==2.3.0",
    "opt-einsum==3.4.0",
    "optax==0.2.5",
    "orbax-checkpoint==0.10.2",
    "packaging==25.0",
    "pandas==2.3.0",
    "pettingzoo==1.25.0",
    "pillow==11.3.0",
    "protobuf==6.31.1",
    "pygame==2.6.1",
    "pygments==2.19.2",
    "pyopengl==3.1.9",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "pytinyrenderer==0.0.14",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "rich==14.0.0",
    "safetensors==0.5.3",
    "scipy==1.12.0",
    "simplejson==3.20.1",
    "six==1.17.0",
    "tensorflow-probability==0.25.0",
    "tensorstore==0.1.75",
    "toolz==1.0.0",
    "tqdm==4.67.1",
    "treescope==0.1.9",
    "trimesh==4.6.13",
    "typing-extensions==4.14.0",
    "tzdata==2025.2",
    "wadler-lindig==0.1.7",
    "werkzeug==3.1.3",
    "wrapt==1.17.2",
    "zipp==3.23.0",
]

[tool.uv]
override-dependencies = ["mujoco==3.2.6", "brax==0.10.4"]
